refactor: Rewrite Google Cloud Storage Health Check as common Health check
